ci: add iOS build and upload to release workflow
Extend the release workflow to build and upload an iOS unsigned IPA package.
This change introduces a new job `build-ios` that compiles the iOS version of
the application and prepares it for distribution. The IPA file is then
uploaded as an artifact to the GitHub release.

The job makes use of environment variables to specify the executable name
and to dynamically create the asset name for the uploaded file. This
enhances the flexibility of the release process for different build targets.